move commands to own file
